Revive joystick code -- more of a re-implementation really.
Configure your joystick through the joystick menu.
Activate the joystick (so it sends events to the UAS) by clicking on the joystick name.

If the UAS is a multi-rotor in stabilize mode, the joystick controls are sent as RC overrides, otherwise if the UAS is in manual mode the joystick controls are sent as manual controls.